Cleaned up all of the code [#158037410]
improved readability
added more comments
removed unused code
improved some function implementations
all of the cleanup all of the time
