+++ projects/revaturecloud/selection/src/lib/components/content-wrapper/content-wrapper.component.ts
@@ -1,5 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 
+/**
+* This component is used to display the room and user card using
+* the lib-user-card-wrapper and lib-room-card-warpper selectors.
+*/
+
 @Component({
   selector: 'lib-content-wrapper',
   templateUrl: './content-wrapper.component.html',
