+++ projects/revaturecloud/selection/src/lib/components/filter-panel/filter-panel.component.ts
@@ -1,10 +1,3 @@
-/**
- * The filter panel of the housing selection front-end.
- *
- * Provides a UI that allows the user to filter and sort
- * the data shown in other components.
- */
-
 import { Component, OnInit } from '@angular/core';
 import { FilterService } from '../../services/filter.service';
 import { FilterSortService } from '../../services/filter-sort.service';
@@ -13,6 +6,13 @@ import { SortParameters } from '../../models/sortParameters.model';
 import { UserStore } from '../../stores/user.store';
 import { RoomStore } from '../../stores/room.store';
 import { Router } from '@angular/router';
+
+/**
+ * The filter panel of the housing selection front-end.
+ *
+ * Provides a UI that allows the user to filter and sort
+ * the data shown in other components.
+ */
 
 @Component({
   // tslint:disable-next-line:component-selector
@@ -58,8 +58,6 @@ export class FilterPanelComponent implements OnInit {
     private filterSortService: FilterSortService,
     private _router: Router) {
 
-    this.reset();
-
     this.filter = {
       batch: null,
       location: null,
@@ -84,23 +82,6 @@ export class FilterPanelComponent implements OnInit {
     this.batches = ['batch1', 'batch2', 'batch3'];
     this.locations = ['Chicago', 'Reston', 'Tampa', 'New York'];
     this.buildings = ['b1', 'b2', 'b3'];
-  }
-
-  /**
-   * Resets the filter form to default values
-   *
-   * DEPRECATED
-   */
-  reset(): void {
-    this.batch = null;
-    this.location = null;
-    this.building = null;
-
-    this.gender = null;
-
-    this.vacantRoomsOnly = false;
-    this.sortByMostVacancies = false;
-    this.unhousedUsersOnly = false;
   }
 
   /**
