+++ projects/revaturecloud/selection/src/lib/components/room-card/room-card.component.html
@@ -13,7 +13,7 @@
         </div>
         <div>
           <font size='2'>{{user.name.last}}</font>
-          <button mat-icon-button (click)="unassign(user, room)">
+          <button mat-icon-button (click)="unassign(user)">
             <mat-icon>remove</mat-icon>
           </button>
         </div>
@@ -26,7 +26,7 @@
         </div>
         <div>
           <font size='2'>{{user.name.last}}</font>
-          <button mat-icon-button (click)="unassign(user, room)">
+          <button mat-icon-button (click)="unassign(user)">
             <mat-icon>remove</mat-icon>
           </button>
         </div>
