+++ projects/revaturecloud/selection/src/lib/components/room-wrapper/room-wrapper.component.ts
@@ -1,10 +1,12 @@
-import { Component, OnInit, ViewChild, OnDestroy, PipeTransform, Pipe } from '@angular/core';
+import { Component, OnInit, OnDestroy} from '@angular/core';
 import { Room } from '../../models/room.model';
-import { Address } from '../../models/address.model';
-import { MatPaginator, PageEvent } from '@angular/material';
-import { Observable } from 'rxjs';
+import { PageEvent } from '@angular/material';
 import { RoomStore } from '../../stores/room.store';
 
+/**
+ * wrapper for displaying room cards with users or unassigned rooms
+ * retrieves data from roomstore and sends corresponding data to room card
+ */
 
 @Component({
   selector: 'lib-room-card-wrapper',
@@ -17,6 +19,7 @@ export class RoomWrapperComponent implements OnInit, OnDestroy {
   searchRooms = '';
   pageEvent: PageEvent;
   pageOptions = [4, 8, 12];
+  p: number = 1;
 
   constructor(private roomStore: RoomStore) {
   }
@@ -28,36 +31,6 @@ export class RoomWrapperComponent implements OnInit, OnDestroy {
       });
   }
 
-  p: number = 1;
-
   ngOnDestroy() {
-  }
-}
-
-@Pipe({
-  name: 'roomSearch',
-  pure: false
-})
-export class RoomSearchPipe implements PipeTransform {
-  transform(value: any, searchString: string): any {
-    if (value.length === 0 || searchString === '') {
-      return value;
-    }
-    searchString = searchString.toLowerCase();
-    const foundRooms = [];
-    for (const room of value) {
-      searchString = searchString.toLowerCase();
-      if (room.location.toLowerCase().includes(searchString)
-        || room.gender.toLowerCase().includes(searchString)
-        // || room.address.address1.toLowerCase().includes(searchString)
-        // || room.address.address2.toLowerCase().includes(searchString)
-        || room.address.city.toLowerCase().includes(searchString)
-        || room.address.state.toLowerCase().includes(searchString)
-        || room.address.postalCode == searchString
-        || room.address.country.toLowerCase().includes(searchString)) {
-        foundRooms.push(room);
-      }
-    }
-    return foundRooms;
   }
 }
