+++ projects/revaturecloud/selection/src/lib/components/table-view/table-view.component.ts
@@ -3,6 +3,11 @@ import { ChangeStore } from '../../stores/change.store';
 import { MatTableDataSource } from '@angular/material';
 import { Tracker } from '../../models/tracker.model';
 
+/**
+* Angular material view of pending room assignment changes
+* retreived data from changestore and displays as table to be confirmed by user
+*/
+
 @Component({
   selector: 'lib-table-view',
   templateUrl: './table-view.component.html',
@@ -10,9 +15,10 @@ import { Tracker } from '../../models/tracker.model';
 })
 export class TableViewComponent implements OnInit {
 
-  constructor(private changeStore: ChangeStore) { }
   displayedColumns = ['User', 'Room', 'Action'];
   dataSource: MatTableDataSource<Tracker>;
+
+  constructor(private changeStore: ChangeStore) { }
 
   ngOnInit() {
     this.changeStore.changes$.subscribe(changes => this.dataSource = new MatTableDataSource(changes));
