+++ projects/revaturecloud/selection/src/lib/components/user-card/user-card.component.ts
@@ -3,6 +3,12 @@ import { User } from '../../models/user.model';
 import { PutService } from '../../services/put.service';
 import { RoomStore } from '../../stores/room.store';
 import { Room } from '../../models/room.model';
+
+/**
+ * The user card component. Displays information about an individual user.
+ * Provides logic for assigned the user to first available
+ */
+
 @Component({
   selector: 'lib-user-card',
   templateUrl: './user-card.component.html',
@@ -12,35 +18,27 @@ export class UserCardComponent implements OnInit {
 
   @Input() user: User;
 
-  constructor(private putService: PutService, private roomStoreService: RoomStore) {}
+  constructor(
+    private putService: PutService,
+    private roomStoreService: RoomStore
+  ) {}
 
   ngOnInit() {}
+
+  /**
+   * Adds this user to the first available room.
+   */
   addToFirstAvailable() {
-    let allRoomsBasedOnCurrentFilter;
-    this.roomStoreService.rooms.subscribe((data: Room[]) => {
-      allRoomsBasedOnCurrentFilter = JSON.parse(JSON.stringify(data));
-    });
-    const allAvailableRoomsBasedOnCurrentFilter: Room[] = allRoomsBasedOnCurrentFilter.filter(
-      (rm: Room) => {
-        if (rm.vacancy) {
-          return (rm.vacancy > 0);
+    this.roomStoreService.rooms.subscribe((res) => {
+      let room: Room = res.find((value) => {
+        if (value.vacancy) {
+          return value.vacancy > 0;
         } else {
-          return false;
+          return false
         }
       });
-    const availableRoom: Room = allAvailableRoomsBasedOnCurrentFilter[0];
-    this.putService.assign(this.user, availableRoom);
-    /**
-     * Debuger info:
-     * console.log('addToFirstAvailable has been called!');
-     * console.log('allRoomsBasedOnCurrentFilter to JSON string:');
-     * console.log(JSON.stringify(allRoomsBasedOnCurrentFilter));
-     * console.log('allAvailableRoomsBasedOnCurrentFilter to JSON string:');
-     * console.log(JSON.stringify(allAvailableRoomsBasedOnCurrentFilter));
-     * console.log('availableRoom to JSON string:');
-     * console.log(JSON.stringify(availableRoom));
-     * console.log('user to JSON string:');
-     */
-  }
 
+      this.putService.assign(this.user, room);
+    });
+  }
 }
