+++ projects/revaturecloud/selection/src/lib/selection.component.html
@@ -1,2 +1,1 @@
 <lib-side-bar></lib-side-bar>
-
