+++ projects/revaturecloud/selection/src/lib/shared/search.pipe.spec.ts
@@ -2,8 +2,6 @@ import { SearchPipe } from './search.pipe';
 
 describe('SearchPipe', () => {
   it('should find the inputs', () => {
-    let searchString = new SearchPipe();
+    //let searchString = new SearchPipe();
   });
 });
-
-
