+++ src/app/app.component.html
@@ -1,6 +1,1 @@
 <lib-selection></lib-selection>
-
-
-
-
-
