+++ src/app/app.module.ts
@@ -1,6 +1,6 @@
 import { NgModule} from '@angular/core';
 import { AppComponent } from './app.component';
-import { SelectionModule } from "../../projects/revaturecloud/selection/src/lib/selection.module";
+import { SelectionModule } from '../../projects/revaturecloud/selection/src/lib/selection.module';
 
 @NgModule({
   declarations: [
